requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.6.3
blinker==1.9.0
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.1
openai==1.59.7
pandas==2.2.3
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
Telethon==1.38.1
tornado==6.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.15
Werkzeug==3.1.3
